pyproject.toml
# Package Name: ewccli
# License: GPL-3.0-or-later
# Copyright (c) 2025 EUMETSAT, ECMWF for European Weather Cloud
# See the LICENSE file for more details

[project]
name = "ewccli"
version = "0.1.1"
description = "Python package to handle EWC resources."
authors = [
    { name = "EWC", email = "support@europeanweather.cloud" }
]
readme = "README.md"
license = { file = "LICENSE" }
maintainers = [
  { name = "EWC", email = "support@europeanweather.cloud"}
]
requires-python = ">3.10"
dependencies = [
    "requests==2.32.5",
    "click==8.3.0",
    "rich==14.1.0",
    "rich-click==1.9.2",
    "prompt_toolkit==3.0.52",
    "pyyaml==6.0.3",
    "cryptography==46.0.2",
    "python-openstackclient==8.2.0",
    "ansible==11.8.0",
    "ansible-runner==2.4.0",
    "kubernetes==34.1.0"
]

classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Topic :: Scientific/Engineering"
]
[project.scripts]
ewc = "ewccli.ewccli:cli"

[project.optional-dependencies]
test = [
    "pytest==8.4.1",
    "pytest-html==4.1.1",
    "pytest-mock==3.14.1",
    "coverage==7.10.5",
    "pre-commit==4.3.0",
    "sphinx==8.1.3",
    "sphinx-click==6.0.0",
    "sphinx-rtd-theme==3.0.2",
    "pydeps==3.0.1",
    "pydantic==2.11.7"
]

[project.urls]
Homepage = "https://www.europeanweather.cloud/"
"Bug Tracker" = "https://github.com/ewcloud/ewccli/issues"
Documentation = "https://github.com/ewcloud/ewccli/blob/main/README.md"
"Source Code" = "https://github.com/ewcloud/ewccli"
Organization = "https://www.europeanweather.cloud/"
"Release Notes" = "https://github.com/ewcloud/ewccli/blob/main/CHANGELOG.md"

[build-system]
requires = ["hatchling>=1.5"]
build-backend = "hatchling.build"

[tool.hatch.envs.default.scripts]
test = "pytest"

[tool.hatch.build.targets.wheel]
packages = ["ewccli"]

[tool.hatch.build.targets.sdist]
include = ["ewccli", "README.md", "LICENSE"]

[tool.coverage.run]
omit = ["*/tests/*"]
